Report.between and .count_between improvements.
Moving the complexity and nastiness out of the helper and specs and into the model while also making the API cleaner and easier to use. Now, Report.between will (as before) give all the reports between two times. But if an interval is given, the return value is an array of reports for each interval. The same goes for Report.count_between.