Adding total change graphs to node index.
This struck some weirdness with Metric.total_changes_between vs. node.metrics.total_changes_between. There needs to be a join to the reports table to get the timestamp, but node.metrics already includes one, so leaving :include => :report in the call to find results in a query with both a left outer join to reports and an inner join to reports. No good. But the include can't be taken away entirely because then Metric.total_changes_between will break.

It's ugly. It hurts.